move faq obj and links arr to 'content' folder
