+++ src/content/index.js
@@ -1,0 +1,2 @@
+export { default as links } from "./links";
+export { default as faq } from "./faq";
