+++ src/content/links.js
@@ -1,0 +1,23 @@
+const links = [
+  {
+    label: "About us",
+    id: "about",
+  },
+  {
+    label: "Services",
+    id: "services",
+  },
+  {
+    label: "Pricing",
+    id: "pricing",
+  },
+  {
+    label: "FAQ",
+    id: "faq",
+  },
+  {
+    label: "Contacts us",
+    id: "contacts",
+  },
+];
+export default links;
